Added login page with react router
