+++ client/src/components/Profile.jsx
@@ -1,10 +1,14 @@
 import React from 'react'
+import { useNavigate } from 'react-router';
+import { Link } from 'react-router-dom';
 
 const Profile = ({profileClass, closeProfile}) => {
+  const navigate = useNavigate();
+
   return (
     <section className={profileClass}>
       <button onClick={closeProfile}>&times;</button>
-      <a href="#">Sign Out</a>
+      <Link to='/login'>Sign Out</Link>
     </section>
   )
 }
