+++ client/src/index.js
@@ -1,10 +1,16 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { createRoot } from 'react-dom/client';
+// import { BrowserRouter } from 'react-router-dom';
 import App from './components/App';
 import './styles.scss';
+import { BrowserRouter } from 'react-router-dom';
 
 // ReactDOM.render(<App />, document.getElementById('root'));
 const container = document.getElementById('root');
 const root = createRoot(container);
-root.render(<App/>);
+root.render(
+  <BrowserRouter>
+    <App/>
+  </BrowserRouter>
+);
