+++ package.json
@@ -27,6 +27,7 @@
     "express": "^4.18.2",
     "mongoose": "^7.6.4",
     "openai": "^4.14.2",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.18.0"
   }
 }
